Added Substantiation Module
Substantiation Module verifies the truthness of a tweet by searching for trusted evidence from the internet. It is implemented by constructing a Deep Self learning BiLSTM NN also known as Siamese Network to capture the sentence similarity. Initially a predicate is created from the tweet. And a Knowledge Graph is created from the content of trusted sources cureated by the Siamese Network by capturing the similarity with the tweet. Then the system checks whether the predicate is present in the Knowledge graph or not and also provides the most similar sentences to substantiate the claim.